fix: purge shadowed docs on flush and release 0.8.1
Clearing shadow markers without rewriting base postings resurrected stale
terms after update+flush. Compact-purge those docs first, and align Python
tests with u64 / NFS2.

+++ test/test_api_coverage.py
@@ -734,7 +734,7 @@ class TestResultHandleMethods:
         arr = result.to_numpy()
         
         assert isinstance(arr, np.ndarray)
-        assert arr.dtype == np.uint32
+        assert arr.dtype == np.uint64
         assert len(arr) == 5
     
     def test_page(self, engine):
